requirements.txt
fastapi
pandas
psycopg2-binary
